fix: create utilities directory and export function
