+++ src/app/page.js
@@ -8,8 +8,7 @@ export default function Home() {
         avatar: "https://ionicframework.com/docs/img/demos/avatar.svg",
         username: "Osgood",
         handle: "@osgoodpaekid",
-        // time: new Date().toISOString().slice(11, 19),
-        // createdAt: new Date().getHours(),
+        createdAt: new Date('2023-07-07'),
         tweet: "reasons why a lot of developers choose next js",
         comment: 50,
         retweet: 31,
